move all logging initialization into initlogging which makes __init__ just a container of metadata (but since we still import the symbols there, any code relying on pysys.ThreadFilter etc will still work just the same); also move logging init code out of pysys.py so it's all together in one place
