+++ pysys/internal/initlogging.py
@@ -17,11 +17,9 @@
 
 # Contains internal (non-API) utilities for initializing logging
 
-import sys, os, io, locale
+import sys, os, io, locale, logging, threading
 
-"""
-@undocumented: PY2, binary_type, _UnicodeSafeStreamWrapper
-"""
+# must not import any pysys packages here, as this module's code needs to execute first
 
 PY2 = sys.version_info[0] == 2
 binary_type = str if PY2 else bytes
@@ -81,3 +79,113 @@ class _UnicodeSafeStreamWrapper(object):
 		self.stream.flush()
 		self.stream.close()
 		self.stream = None
+
+
+# class extensions for supporting multi-threaded nature
+
+class ThreadedStreamHandler(logging.StreamHandler):
+	"""Stream handler to only log from the creating thread.
+	
+	Overrides logging.StreamHandler to only allow logging to a stream 
+	from the thread that created the class instance and added to the root 
+	logger via log.addHandler(ThreadedStreamHandler(stream)).
+
+	This is used to pass log output from the specific test that creates this 
+	handler to stdout, either immediately or (when multiple threads are in use) 
+	at the end of each test's execution. 
+	
+	@deprecated: For internal use only, do not use. 
+	"""
+	def __init__(self, strm=None, streamFactory=None):
+		"""Overrides logging.StreamHandler.__init__.
+		@param strm: the stream
+		@param streamFactory: a function that returns the stream, if strm is not specified. 
+		"""
+		self.threadId = threading.current_thread().ident
+		self.__streamfactory = streamFactory if streamFactory else (lambda:strm)
+		logging.StreamHandler.__init__(self, self.__streamfactory())
+		
+	def emit(self, record):
+		"""Overrides logging.StreamHandler.emit."""
+		if self.threadId != threading.current_thread().ident: return
+		logging.StreamHandler.emit(self, record)
+
+	def _updateUnderlyingStream(self):
+		""" Update the stream this handler uses by calling again the stream factory; 
+		used only for testing. 
+		"""
+		assert self.stream # otherwise assigning to it wouldn't do anything
+		self.stream = self.__streamfactory()
+
+class ThreadedFileHandler(logging.FileHandler):
+	"""File handler to only log from the creating thread.
+	
+	Overrides logging.FileHandler to only allow logging to file from 
+	the thread than created the class instance and added to the root 
+	logger via log.addHandler(ThreadFileHandler(filename)).
+	
+	This is used to pass log output from the specific test 
+	that creates this handler to the associated run.log. 
+
+	@deprecated: No longer used, will be removed. 
+	"""
+	def __init__(self, filename, encoding=None):
+		"""Overrides logging.FileHandler.__init__"""
+		self.threadId = threading.current_thread().ident
+		logging.FileHandler.__init__(self, filename, "a", encoding=self.__streamencoding)
+
+	def emit(self, record):
+		"""Overrides logging.FileHandler.emit."""
+		if self.threadId != threading.current_thread().ident: return
+		# must put formatted messages into the buffer otherwise we lose log level 
+		# and (critically) exception tracebacks from the output
+		logging.FileHandler.emit(self, record)
+		
+
+class ThreadFilter(logging.Filterer):
+	"""Filter to disallow log records from the current thread.
+	
+	Within pysys, logging to standard output is only enabled from the main thread 
+	of execution (that in which the test runner class executes). When running with
+	more than one test worker thread, logging to file of the test run log is 
+	performed through a file handler, which only allows logging from that thread. 
+	To disable either of these, use an instance of this class from the thread in 
+	question, adding to the root logger via log.addFilter(ThreadFilter()).
+	
+	"""
+	def __init__(self):
+		"""Overrides logging.Filterer.__init__"""
+		self.threadId = threading.current_thread().ident
+		logging.Filterer.__init__(self)
+		
+	def filter(self, record):
+		"""Implementation of logging.Filterer.filter to block from the creating thread."""
+		if self.threadId != threading.current_thread().ident: return True
+		return False
+
+
+#####################################
+
+# Initialize Python logging for PySys
+
+# avoids a bug where error handlers using the Python root handler could mess up 
+# subsequent logging if no handlers are defined
+logging.getLogger().addHandler(logging.NullHandler())
+
+rootLogger = logging.getLogger('pysys')
+"""The root logger for logging within PySys."""
+
+log = rootLogger
+
+stdoutHandler = ThreadedStreamHandler(streamFactory=lambda: _UnicodeSafeStreamWrapper(sys.stdout, writebytes=PY2))
+"""The handler that sends pysys.* log output from the main thread to stdout, 
+including buffered output from completed tests when running in parallel."""
+
+
+# customize the default logging names for display
+logging.addLevelName(50, 'CRIT')
+logging.addLevelName(30, 'WARN')
+stdoutHandler.setLevel(logging.INFO)
+rootLogger.setLevel(logging.DEBUG) # The root logger log level (set to DEBUG as all filtering is done by the handlers).
+rootLogger.addHandler(stdoutHandler)
+
